requirements.txt
alpaca-py>=0.40.0
anthropic>=0.25.0
aiosqlite>=0.20.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
